Split video watch page sections
